MM-19: Separate controllers
