--- client/controller.py
@@ -1,376 +1,0 @@
-import socket as sock
-import json
-import sys
-from threading import Thread
-from view import *
-import curses
-from typing import *
-import curses.ascii
-import traceback
-
-
-class Controller:
-    def __init__(self):
-        self.view: View = View(self)
-
-    def start(self):
-        curses.wrapper(self.view.display)
-
-    def get_input(self) -> None:
-        pass
-
-
-class Menu(Controller):
-    def __init__(self, menu):
-        super().__init__()
-        self.menu = menu
-        self.cursor: int = 0
-        self.view = MenuView(self)
-
-    def get_input(self) -> int:
-        key = self.view.stdscr.getch()
-
-        # Movement
-        if key == curses.KEY_UP:
-            self.cursor = max(self.cursor - 1, 0)
-        elif key == curses.KEY_DOWN:
-            self.cursor = min(self.cursor + 1, len(self.menu) - 1)
-        elif key == curses.ascii.TAB:
-            self.cursor = (self.cursor + 1) % len(self.menu)
-        elif key in (curses.ascii.LF, curses.ascii.CR, curses.ascii.BEL, curses.KEY_ENTER):
-            fn = self.menu[list(self.menu.keys())[self.cursor]]
-            if fn is not None:
-                fn()
-        elif key == ord('q'):
-            self.view.stop()
-
-        return key
-
-
-class NetworkMenu(Menu):
-    """
-    A menu which requires a connection to the server to send and receive
-    packets. Examples include a login menu or a registration menu.
-    """
-    def __init__(self, host: str, port: int, menu: dict):
-        super().__init__(menu)
-
-        self.addr = (host, port)
-        self.s = None
-
-    """
-    Attempts to connect to the network and start the menu. Returns an error
-    message if unsuccessful.
-    """
-    def start(self, attempt=0) -> str:
-        max_attempts: int = 3
-        self.s: sock.socket = sock.socket(sock.AF_INET, sock.SOCK_STREAM)
-        try:
-            self.s.connect(self.addr)
-        except sock.error as e:
-            if attempt < max_attempts:
-                # Socket might have closed, open it again
-                self.start(attempt=attempt+1)
-                print(f"Error: Connection attempt {attempt}/{max_attempts} failed. Traceback: ", file=sys.stderr)
-                print(traceback.format_exc(), file=sys.stderr)
-            else:
-                # Something else is wrong, exit
-                print(f"Error: Connection attempts reached max ({max_attempts}). Trackback: ", file=sys.stderr)
-                print(traceback.format_exc(), file=sys.stderr)
-                self.view.stop()
-                return str(e)
-
-        super().start()
-
-
-class MainMenu(Menu):
-    def __init__(self, host, port):
-        super().__init__({
-            "Login": self.login,
-            "Register": self.register
-        })
-
-        self.host = host
-        self.port = port
-
-        self.view = MenuView(self, f"Welcome to {host}:{port}")
-
-    def login(self):
-        loginmenu = LoginMenu(self.host, self.port)
-        err: str = loginmenu.start()
-        if err:
-            self.view.title = err
-            return
-        self.start()
-
-    def register(self):
-        registermenu = RegisterMenu(self.host, self.port)
-        err: str = registermenu.start()
-        if err:
-            self.view.title = err
-            return
-        self.start()
-
-
-class LoginMenu(NetworkMenu):
-    def __init__(self, host, port):
-        self.username: str = ''
-        self.password: str = ''
-
-        super().__init__(host, port, {
-            "Username": self.login,
-            "Password": self.login
-        })
-
-        self.view = LoginView(self)
-
-    def get_input(self) -> int:
-        key = super().get_input()
-        if curses.ascii.isprint(key) or key in (curses.KEY_LEFT, curses.KEY_RIGHT, curses.KEY_DC):
-            if self.cursor == 0:
-                self.view.usernamebox.modal(first_key=key)
-                self.username = self.view.usernamebox.value
-            elif self.cursor == 1:
-                self.view.passwordbox.modal(first_key=key)
-                self.password = self.view.passwordbox.value
-        return key
-
-    def login(self):
-        if '' in (self.username, self.password):
-            self.view.title = "Username or password must not be blank"
-            return
-
-        self.view.title = f"Attempted login in as {self.username} with password {self.password}"
-        try:
-            self.s.send(bytes(json.dumps({
-                'a': 'login',
-                'p': self.username,
-                'p2': self.password
-            }) + ';', 'utf-8'))
-
-            game = Game(self.s, self.addr)
-            game.start()
-            self.start()
-
-        except sock.error as e:
-            print("Error: Socket error. Traceback: ", file=sys.stderr)
-            print(traceback.format_exc())
-            self.view.title = str(e)
-
-
-class RegisterMenu(NetworkMenu):
-    def __init__(self, host, port):
-        self.username: str = ''
-        self.password: str = ''
-        self.confirmpassword: str = ''
-
-        super().__init__(host, port, {
-            "Username": self.register,
-            "Password": self.register,
-            "Confirm password": self.register
-        })
-
-        self.view = RegisterView(self)
-
-    def get_input(self) -> int:
-        key = super().get_input()
-        if curses.ascii.isprint(key) or key in (curses.KEY_LEFT, curses.KEY_RIGHT, curses.KEY_DC):
-            if self.cursor == 0:
-                self.view.usernamebox.modal(first_key=key)
-                self.username = self.view.usernamebox.value
-            elif self.cursor == 1:
-                self.view.passwordbox.modal(first_key=key)
-                self.password = self.view.passwordbox.value
-            elif self.cursor == 2:
-                self.view.confirmpasswordbox.modal(first_key=key)
-                self.confirmpassword = self.view.confirmpasswordbox.value
-        return key
-
-    def register(self):
-        if '' in (self.username, self.password, self.confirmpassword):
-            self.view.title = "Field must not be blank"
-            return
-
-        self.view.title = f"Attempted registration as {self.username} with password {self.password}"
-
-        if self.password != self.confirmpassword:
-            self.view.title = "Passwords do not match!"
-            return
-
-        try:
-            self.s.send(bytes(json.dumps({
-                'a': 'register',
-                'p': self.username,
-                'p2': self.password
-            }) + ';', 'utf-8'))
-
-        except sock.error as e:
-            print("Error: Socket error. Traceback: ", file=sys.stderr)
-            print(traceback.format_exc())
-            self.view.title = str(e)
-
-
-class Game(Controller):
-    def __init__(self, s: sock.socket, addr: Tuple[str, int]):
-        super().__init__()
-        self.s: sock.socket = s
-        self.addr: Tuple[str, int] = addr
-
-        self.connected = False
-
-        self.player_id: int = 0
-        self.game_data: dict = {}
-        self.walls: list = []
-
-        self.size: Tuple[int, int] = (-1, -1)
-        self.tick_rate: int = -1
-
-        # UI
-        self.chatbox = None
-        self.view = GameView(self)
-
-    def connect(self) -> None:
-        try:
-            self.s.connect(self.addr)
-        except sock.error:
-            # Most likely safe to ignore because socket should be already connected
-            print(f"Error: Connection refused. Traceback: ", file=sys.stderr)
-            print(traceback.format_exc(), file=sys.stderr)
-        message: str = ""
-        while True:
-            try:
-                message += self.s.recv(1).decode('utf-8')
-            except sock.error as e:
-                print("Error: Error receiving message. Traceback: ", file=sys.stderr)
-                print(traceback.format_exc())
-                self.chat("I got an error receiving a message!")
-
-            if message[-1] == ";":
-                if message[:-1] == 'full':
-                    print('Session is full.')
-                    sys.exit()
-                else:
-                    data = json.loads(message[:-1])
-                    break
-
-        self.size = (data['h'], data['w'])
-        self.player_id = data['id']
-        self.walls = data['walls']
-        self.tick_rate = data['t']
-
-        self.connected = True
-
-    def disconnect(self) -> None:
-        try:
-            # Action: move, Payload: direction
-            self.s.send(bytes(json.dumps({
-                'a': 'bye'
-            }) + ';', 'utf-8'))
-        except sock.error as e:
-            print("Error: Socket error. Traceback: ", file=sys.stderr)
-            print(traceback.format_exc())
-            print(e, file=sys.stderr)
-        self.s.close()
-
-        self.connected = False
-
-    def start(self) -> None:
-        self.connect()
-
-        # Listen for game data in its own thread
-        Thread(target=self.load_data, daemon=True).start()
-
-        # Don't use game data until it's been received
-        while self.game_data == {}:
-            time.sleep(0.2)
-
-        super().start()
-
-    def load_data(self) -> None:
-        message = ""
-        while self.connected:
-            try:
-                message += self.s.recv(1).decode('utf-8')
-                if message[-1] == ";":
-                    if message[-2] != '\\':
-                        # Terminate, end of JSON clause
-                        self.game_data = json.loads(message[:-1])
-                    else:
-                        # Interpret as a literal
-                        message = message[:-3] + ';'
-                        continue
-                    message = ""
-            except sock.error:
-                print("Error: Socket error. Traceback: ", file=sys.stderr)
-                print(traceback.format_exc())
-                message = ""
-
-    def get_input(self) -> None:
-        try:
-            key = self.view.stdscr.getch()
-
-            # Movement
-            if key == curses.KEY_UP:
-                self.move(0)
-            elif key == curses.KEY_RIGHT:
-                self.move(1)
-            elif key == curses.KEY_DOWN:
-                self.move(2)
-            elif key == curses.KEY_LEFT:
-                self.move(3)
-
-            # Changing window focus
-            elif key in (ord('1'), ord('2'), ord('3')):
-                self.view.focus = int(chr(key))
-
-            # Changing window 2 focus
-            elif key == ord('?'):
-                self.view.win2_focus = Window2Focus.HELP
-            elif key == ord('k'):
-                self.view.win2_focus = Window2Focus.SKILLS
-            elif key == ord('i'):
-                self.view.win2_focus = Window2Focus.INVENTORY
-            elif key == ord('p'):
-                self.view.win2_focus = Window2Focus.SPELLBOOK
-            elif key == ord('g'):
-                self.view.win2_focus = Window2Focus.GUILD
-            elif key == ord('j'):
-                self.view.win2_focus = Window2Focus.JOURNAL
-
-            # Chat
-            elif key in (curses.KEY_ENTER, curses.ascii.LF, curses.ascii.CR) and self.view.chatbox is not None:
-                self.view.chatbox.modal()
-                self.chat(self.view.chatbox.value)
-
-            elif key == ord('q'):
-                self.disconnect()
-                self.view.stop()
-
-        except KeyboardInterrupt:
-           self.disconnect()
-           exit()
-
-    def move(self, direction: int) -> None:
-        try:
-            # Action: move, Payload: direction
-            self.s.send(bytes(json.dumps({
-                'a': 'm',
-                'p': direction
-            }) + ';', 'utf-8'))
-        except sock.error:
-            print("Error: Socket error. Traceback: ", file=sys.stderr)
-            print(traceback.format_exc())
-
-    def chat(self, message: str) -> None:
-        # Sanitise semicolons
-        message = message.replace(';', '\\;')
-
-        try:
-            # Action: chat, Payload: message
-            self.s.send(bytes(json.dumps({
-                'a': 'c',
-                'p': message
-            }) + ';', 'utf-8'))
-        except sock.error:
-            print("Error: Socket error. Traceback: ", file=sys.stderr)
-            print(traceback.format_exc())
