+++ client/controllers/__init__.py
@@ -1,0 +1,7 @@
+from .controller import Controller
+from .menu import Menu
+from .networkmenu import NetworkMenu
+from .mainmenu import MainMenu
+from .loginmenu import LoginMenu
+from .registermenu import RegisterMenu
+from .game import Game
